Adds filesystem fallback search for kernel headers
When building in a container-based build service, `uname -r` may not retrieve the kernel headers installed on the OS. This adds a fallback option to taking the first set of headers found in the `/usr/src/kernels` directory. If headers for the current running kernel are found first, these will be used instead.