RX_Ring: Remove extra blocks data structure
Okay, this might be naive. Both Suricata and the AF_PACKET example [1]
keep the block descriptor offsets in a separately allocated table as
well.

From all I can tell, that allocation isn't really needed if we're
okay computing the address of the block based on the beginning of the
mmap'ed memory region using the tp_req parameters. That shouldn't
be slower

[1] https://docs.kernel.org/networking/packet_mmap.html#af-packet-tpacket-v3-example