Drop current Github CI workflows
With this plugin built-in to Zeek by default, nightly tests have been failing
for a while now. The tests technically still work for the LTS binaries, but
those too will soon fail when 6.0 comes out.

Going forward it'll be better to test this package's development with a Zeek
image that does not have the plugin built in.